+++ processor/src/test/resources/bad_input/com/example/component/UnexpectedAbstractComponent.java
@@ -1,0 +1,8 @@
+package com.example.component;
+
+import arez.annotations.ArezComponent;
+
+@ArezComponent( allowConcrete = true )
+public abstract class UnexpectedAbstractComponent
+{
+}